refactored code moved common maps services and types to shared
