+++ src/features/dashboard/sections/DashboardMaps/types/map.ts
@@ -1,14 +1,1 @@
-export interface Map {
-  readonly id: string
-  readonly name: string
-  manifestId: string
-  description?: string
-  isTemplate: boolean
-  forkedFrom?: string
-  createdBy: string
-  isPublic: boolean
-  category?: string
-  previewImageUrl: string
-  createdAt: Date
-  updatedAt: Date
-}
+export * from '@/shared/types'
