+++ src/shared/schema/index.ts
@@ -1,3 +1,6 @@
 export * from './common'
 export * from './common.schema'
 export * from './validation.schemas'
+export * from './manifest/manifest.scema'
+export * from './map/map.scema'
+export * from './upload/upload.schema'
