+++ src/shared/schema/manifest/manifest.scema.ts
