+++ src/shared/schema/map/map.scema.ts
@@ -1,5 +1,5 @@
 import z from 'zod'
-import { ManifestData } from './manifest.schema'
+import { ManifestData } from '../manifest/manifest.scema'
 
 const MAX_FILE_SIZE = 5000000
 const ACCEPTED_IMAGE_TYPES = ['image/jpeg', 'image/jpg', 'image/png', 'image/webp']
