+++ src/shared/services/index.ts
@@ -1,0 +1,2 @@
+export * from './map'
+export * from './upload'
